requirements.txt
opencv-python
torch
torchvision
Pillow
numpy
timm